refactor(convert): move pseudo.rs tests into home module
