+++ crates/fulgur/src/convert/mod.rs
@@ -1426,7 +1426,7 @@ mod tests {
         Arc::new(TEST_PNG_1X1.to_vec())
     }
 
-    fn find_h1(doc: &blitz_html::HtmlDocument) -> usize {
+    pub(super) fn find_h1(doc: &blitz_html::HtmlDocument) -> usize {
         fn walk(doc: &blitz_dom::BaseDocument, id: usize) -> Option<usize> {
             let node = doc.get_node(id)?;
             if let Some(ed) = node.element_data() {
@@ -1444,150 +1444,6 @@ mod tests {
         walk(doc.deref(), doc.root_element().id).expect("h1 not found")
     }
 
-    #[test]
-    fn test_build_pseudo_image_reads_content_url() {
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .expect("read examples/image/icon.png");
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            h1::before {
-                content: url("icon.png");
-                display: block;
-                width: 48px;
-                height: 48px;
-            }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-
-        let h1_id = find_h1(&doc);
-        let before_id = doc
-            .get_node(h1_id)
-            .unwrap()
-            .before
-            .expect("::before pseudo");
-        let pseudo = doc.get_node(before_id).unwrap();
-        let (parent_w, parent_h) = size_in_pt(doc.get_node(h1_id).unwrap().final_layout.size);
-
-        let img = super::pseudo::build_pseudo_image(pseudo, parent_w, parent_h, Some(&bundle))
-            .expect("build_pseudo_image should return Some for content: url()");
-        // 48 CSS px × 0.75 = 36 pt
-        assert_eq!(img.width, 36.0);
-        assert_eq!(img.height, 36.0);
-    }
-
-    #[test]
-    fn test_build_pseudo_image_width_only_uses_intrinsic_aspect() {
-        // icon.png is 32x32 so aspect = 1.0. width:20px → 15 pt, height
-        // back-propagates via intrinsic aspect → 15 pt.
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("icon.png"); display: block; width: 20px; }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.unwrap();
-        let pseudo = doc.get_node(before_id).unwrap();
-        let (parent_w, parent_h) = size_in_pt(doc.get_node(h1_id).unwrap().final_layout.size);
-
-        let img =
-            super::pseudo::build_pseudo_image(pseudo, parent_w, parent_h, Some(&bundle)).unwrap();
-        assert_eq!(img.width, 15.0);
-        assert_eq!(img.height, 15.0);
-    }
-
-    #[test]
-    fn test_build_pseudo_image_missing_asset_returns_none() {
-        let bundle = AssetBundle::new();
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("missing.png"); display: block; }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.unwrap();
-        let pseudo = doc.get_node(before_id).unwrap();
-        assert!(
-            super::pseudo::build_pseudo_image(pseudo, 800.0, 600.0, Some(&bundle)).is_none(),
-            "missing asset should silently return None"
-        );
-    }
-
-    #[test]
-    fn test_build_pseudo_image_no_assets_returns_none() {
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("icon.png"); }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.unwrap();
-        let pseudo = doc.get_node(before_id).unwrap();
-        assert!(super::pseudo::build_pseudo_image(pseudo, 800.0, 600.0, None).is_none());
-    }
-
-    #[test]
-    fn test_build_pseudo_image_height_percent_resolves_against_parent_height() {
-        // Verifies the coderabbit fix: height: 50% on the pseudo should
-        // resolve against parent_content_height, not parent_content_width.
-        // icon.png is 32x32 intrinsic, so with height=50% of 200 = 100 and
-        // no explicit width, the aspect ratio gives width = 100.
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("icon.png"); display: block; height: 50%; }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.unwrap();
-        let pseudo = doc.get_node(before_id).unwrap();
-
-        // Explicitly call with distinguishable width (400) and height (200)
-        // so we can verify which basis is used for `height: 50%`.
-        let img = super::pseudo::build_pseudo_image(pseudo, 400.0, 200.0, Some(&bundle)).unwrap();
-        assert_eq!(
-            img.height, 100.0,
-            "height: 50% should resolve against parent_content_height (200.0)"
-        );
-        assert_eq!(
-            img.width, 100.0,
-            "intrinsic aspect (1:1) should give width = height"
-        );
-    }
-
     /// Recursively walk a Pageable tree and push any ImagePageable found.
     pub(super) fn collect_images(p: &dyn Pageable, out: &mut Vec<(f32, f32)>) {
         if let Some(img) = p.as_any().downcast_ref::<ImagePageable>() {
@@ -1601,261 +1457,10 @@ mod tests {
         }
     }
 
-    #[test]
-    fn test_dom_to_pageable_emits_block_pseudo_image() {
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            .wrap::before {
-                content: url("icon.png");
-                display: block;
-                width: 24px;
-                height: 24px;
-            }
-        </style></head><body><div class="wrap">hello</div></body></html>"#;
-
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-
-        let running_store = crate::gcpm::running::RunningElementStore::new();
-        let mut ctx = ConvertContext {
-            running_store: &running_store,
-            assets: Some(&bundle),
-            font_cache: HashMap::new(),
-            string_set_by_node: HashMap::new(),
-            counter_ops_by_node: HashMap::new(),
-            bookmark_by_node: HashMap::new(),
-            column_styles: crate::column_css::ColumnStyleTable::new(),
-            multicol_geometry: crate::multicol_layout::MulticolGeometryTable::new(),
-            link_cache: Default::default(),
-        };
-        let tree = super::dom_to_pageable(&doc, &mut ctx);
-
-        let mut images = Vec::new();
-        collect_images(&*tree, &mut images);
-        assert!(
-            images.iter().any(|(w, h)| *w == 18.0 && *h == 18.0),
-            "expected an 18x18 pt ImagePageable (24 CSS px × 0.75) from ::before pseudo, got {:?}",
-            images
-        );
-    }
-
-    #[test]
-    fn test_dom_to_pageable_inline_pseudo_ignored_phase1() {
-        // Phase 1 only handles display:block pseudos. An inline pseudo with
-        // content: url() should be silently ignored (Phase 2 will handle it).
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            p::before { content: url("icon.png"); width: 10px; height: 10px; }
-        </style></head><body><p>hello</p></body></html>"#;
-
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-
-        let running_store = crate::gcpm::running::RunningElementStore::new();
-        let mut ctx = ConvertContext {
-            running_store: &running_store,
-            assets: Some(&bundle),
-            font_cache: HashMap::new(),
-            string_set_by_node: HashMap::new(),
-            counter_ops_by_node: HashMap::new(),
-            bookmark_by_node: HashMap::new(),
-            column_styles: crate::column_css::ColumnStyleTable::new(),
-            multicol_geometry: crate::multicol_layout::MulticolGeometryTable::new(),
-            link_cache: Default::default(),
-        };
-        let tree = super::dom_to_pageable(&doc, &mut ctx);
-
-        let mut images = Vec::new();
-        collect_images(&*tree, &mut images);
-        assert!(
-            images.is_empty(),
-            "Phase 1 must not emit inline pseudo images; got {:?}",
-            images
-        );
-    }
-
-    #[test]
-    fn test_dom_to_pageable_emits_pseudo_on_childless_element() {
-        // Regression for Devin Review comment on PR #70: the children.is_empty()
-        // branch used to skip pseudo injection. `<div class="icon"></div>` with
-        // a block pseudo should still render the image.
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            .icon::before {
-                content: url("icon.png");
-                display: block;
-                width: 16px;
-                height: 16px;
-            }
-        </style></head><body><div class="icon"></div></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let running_store = crate::gcpm::running::RunningElementStore::new();
-        let mut ctx = ConvertContext {
-            running_store: &running_store,
-            assets: Some(&bundle),
-            font_cache: HashMap::new(),
-            string_set_by_node: HashMap::new(),
-            counter_ops_by_node: HashMap::new(),
-            bookmark_by_node: HashMap::new(),
-            column_styles: crate::column_css::ColumnStyleTable::new(),
-            multicol_geometry: crate::multicol_layout::MulticolGeometryTable::new(),
-            link_cache: Default::default(),
-        };
-        let tree = super::dom_to_pageable(&doc, &mut ctx);
-        let mut images = Vec::new();
-        collect_images(&*tree, &mut images);
-        assert!(
-            images.iter().any(|(w, h)| *w == 12.0 && *h == 12.0),
-            "childless element ::before pseudo should emit a 12x12 pt image (16 CSS px × 0.75); got {:?}",
-            images
-        );
-    }
-
-    #[test]
-    fn test_dom_to_pageable_emits_pseudo_on_zero_size_block_leaf() {
-        // Regression for coderabbit follow-up on PR #70: a 0x0 block leaf
-        // was being skipped by the collect_positioned_children zero-size
-        // leaf filter BEFORE reaching the convert_node `children.is_empty()`
-        // branch. The pseudo probe (`node_has_block_pseudo_image`) now lets
-        // such leaves fall through.
-        //
-        // Scope note: this test specifically targets a BLOCK element with
-        // explicit width:0;height:0 that still has a ::before image — e.g.
-        // a decorative sentinel div a template sets to 0x0 with a pseudo
-        // icon. Inline `<span>` with a block ::before is a different
-        // edge case that requires Phase 2 inline-flow handling.
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            .zero { display: block; width: 0; height: 0; }
-            .zero::before {
-                content: url("icon.png");
-                display: block;
-                width: 18px;
-                height: 18px;
-            }
-        </style></head><body><section><div class="zero"></div></section></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let running_store = crate::gcpm::running::RunningElementStore::new();
-        let mut ctx = ConvertContext {
-            running_store: &running_store,
-            assets: Some(&bundle),
-            font_cache: HashMap::new(),
-            string_set_by_node: HashMap::new(),
-            counter_ops_by_node: HashMap::new(),
-            bookmark_by_node: HashMap::new(),
-            column_styles: crate::column_css::ColumnStyleTable::new(),
-            multicol_geometry: crate::multicol_layout::MulticolGeometryTable::new(),
-            link_cache: Default::default(),
-        };
-        let tree = super::dom_to_pageable(&doc, &mut ctx);
-        let mut images = Vec::new();
-        walk_all_children(&*tree, &mut |p| collect_images(p, &mut images));
-        assert!(
-            images.iter().any(|(w, h)| *w == 13.5 && *h == 13.5),
-            "zero-size block leaf with block pseudo should emit a 13.5x13.5 pt image (18 CSS px × 0.75); got {:?}",
-            images
-        );
-    }
-
-    #[test]
-    fn test_dom_to_pageable_emits_pseudo_on_list_item_with_text() {
-        // Regression for Devin Review comment on PR #70: the list item
-        // inline-root body path used to skip pseudo injection. A <li> with
-        // inline text content and a block ::before should still render.
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .unwrap();
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            li::before {
-                content: url("icon.png");
-                display: block;
-                width: 12px;
-                height: 12px;
-            }
-        </style></head><body><ul><li>item text</li></ul></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let running_store = crate::gcpm::running::RunningElementStore::new();
-        let mut ctx = ConvertContext {
-            running_store: &running_store,
-            assets: Some(&bundle),
-            font_cache: HashMap::new(),
-            string_set_by_node: HashMap::new(),
-            counter_ops_by_node: HashMap::new(),
-            bookmark_by_node: HashMap::new(),
-            column_styles: crate::column_css::ColumnStyleTable::new(),
-            multicol_geometry: crate::multicol_layout::MulticolGeometryTable::new(),
-            link_cache: Default::default(),
-        };
-        let tree = super::dom_to_pageable(&doc, &mut ctx);
-        let mut images = Vec::new();
-        walk_all_children(&*tree, &mut |p| collect_images(p, &mut images));
-        assert!(
-            images.iter().any(|(w, h)| *w == 9.0 && *h == 9.0),
-            "list item with text + block pseudo should emit a 9x9 pt image (12 CSS px × 0.75); got {:?}",
-            images
-        );
-    }
-
     /// Walk a Pageable tree visiting every nested child via known container
     /// types. Used by tests that need to peek through `ListItemPageable`'s
     /// body, which the simple BlockPageable-only walker does not descend into.
-    fn walk_all_children(p: &dyn Pageable, visit: &mut dyn FnMut(&dyn Pageable)) {
+    pub(super) fn walk_all_children(p: &dyn Pageable, visit: &mut dyn FnMut(&dyn Pageable)) {
         visit(p);
         if let Some(block) = p.as_any().downcast_ref::<BlockPageable>() {
             for c in &block.children {
@@ -1991,85 +1596,6 @@ mod tests {
                 i.x_offset,
             );
         }
-    }
-
-    #[test]
-    fn test_build_inline_pseudo_image_returns_some_for_inline_pseudo() {
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .expect("icon.png fixture must exist");
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("icon.png"); width: 24px; height: 24px; }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.expect("::before");
-        let pseudo = doc.get_node(before_id).unwrap();
-
-        // Inline pseudos have display: inline by default (not block)
-        assert!(
-            !super::pseudo::is_block_pseudo(pseudo),
-            "pseudo should be inline by default"
-        );
-
-        let img = super::pseudo::build_inline_pseudo_image(pseudo, 800.0, 600.0, Some(&bundle));
-        assert!(img.is_some(), "should return Some for inline pseudo");
-        let img = img.unwrap();
-        // 24 CSS px × 0.75 = 18 pt
-        assert_eq!(img.width, 18.0);
-        assert_eq!(img.height, 18.0);
-    }
-
-    #[test]
-    fn test_build_inline_pseudo_image_does_not_filter_display() {
-        let icon_bytes = std::fs::read(
-            std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-                .parent()
-                .unwrap()
-                .parent()
-                .unwrap()
-                .join("examples/image/icon.png"),
-        )
-        .expect("icon.png fixture must exist");
-        let mut bundle = AssetBundle::new();
-        bundle.add_image("icon.png", icon_bytes);
-
-        let html = r#"<!doctype html><html><head><style>
-            h1::before { content: url("icon.png"); display: block; width: 24px; height: 24px; }
-        </style></head><body><h1>T</h1></body></html>"#;
-        let mut doc = crate::blitz_adapter::parse(html, 800.0, &[]);
-        crate::blitz_adapter::resolve(&mut doc);
-        let h1_id = find_h1(&doc);
-        let before_id = doc.get_node(h1_id).unwrap().before.expect("::before");
-        let pseudo = doc.get_node(before_id).unwrap();
-
-        assert!(
-            super::pseudo::is_block_pseudo(pseudo),
-            "pseudo with display:block should be block"
-        );
-
-        // The inline builder should NOT produce an image for block pseudos
-        // (the caller filters with !is_block_pseudo, but we verify the function
-        // itself still returns Some — the filtering is done at the call site)
-        // Here we verify the function works, the call-site filter is tested
-        // by the integration test above.
-        let img = super::pseudo::build_inline_pseudo_image(pseudo, 800.0, 600.0, Some(&bundle));
-        // build_inline_pseudo_image doesn't check display, so this will be Some.
-        // The call site filters with !is_block_pseudo.
-        assert!(
-            img.is_some(),
-            "build_inline_pseudo_image itself doesn't filter display"
-        );
     }
 
     /// Walk the DOM to find the first element with `tag` and return its node id.
